Revamped cache, using cachetools instead of fastapi-cache2
Use cache only for Modrinth and Hangar interfaces. API endpoints are not
cached. As we already recommend a reverse-proxy for authentication, we
might as well use it for caching. API endpoints do set the proper cache
headers so that no configuration, either on the reverse-proxy or on the
client, should be necessary.